+++ plugins/destination/clickhouse/queries/read_test.go
@@ -4,7 +4,8 @@ import (
 	"testing"
 
 	"github.com/ClickHouse/clickhouse-go/v2/lib/driver"
-	"github.com/cloudquery/plugin-sdk/v2/schema"
+	"github.com/apache/arrow/go/v13/arrow"
+	"github.com/cloudquery/plugin-sdk/v3/schema"
 	"github.com/stretchr/testify/require"
 )
 
@@ -13,16 +14,16 @@ func TestRead(t *testing.T) {
 		sourceName = "cq_source"
 	)
 
-	query, params := Read(sourceName, schema.CQSchemaToArrow(&schema.Table{
+	query, params := Read(sourceName, &schema.Table{
 		Name: "table_name",
 		Columns: schema.ColumnList{
 			schema.CqIDColumn,
 			schema.CqParentIDColumn,
 			schema.CqSourceNameColumn,
 			schema.CqSyncTimeColumn,
-			schema.Column{Name: "extra_col", Type: schema.TypeFloat},
+			schema.Column{Name: "extra_col", Type: arrow.PrimitiveTypes.Float64},
 		},
-	}))
+	})
 
 	ensureContents(t, query, "read.sql")
 	require.Equal(t, 1, len(params))