+++ plugins/destination/clickhouse/queries/tables.go
@@ -3,18 +3,17 @@ package queries
 import (
 	"strings"
 
-	"github.com/apache/arrow/go/v13/arrow"
 	"github.com/cloudquery/cloudquery/plugins/destination/clickhouse/typeconv/ch/types"
 	"github.com/cloudquery/cloudquery/plugins/destination/clickhouse/util"
-	"github.com/cloudquery/plugin-sdk/v2/schema"
+	"github.com/cloudquery/plugin-sdk/v3/schema"
 	"golang.org/x/exp/slices"
 )
 
-func sortKeys(sc *arrow.Schema) []string {
-	keys := make([]string, 0, len(sc.Fields()))
-	for _, field := range sc.Fields() {
-		if !field.Nullable {
-			keys = append(keys, field.Name)
+func sortKeys(table *schema.Table) []string {
+	keys := make([]string, 0, len(table.Columns))
+	for _, col := range table.Columns {
+		if col.NotNull {
+			keys = append(keys, col.Name)
 		}
 	}
 
@@ -27,26 +26,25 @@ func sortKeys(sc *arrow.Schema) []string {
 	return slices.Clip(keys)
 }
 
-func CreateTable(sc *arrow.Schema, cluster string, engine *Engine) (string, error) {
+func CreateTable(table *schema.Table, cluster string, engine *Engine) (string, error) {
 	builder := strings.Builder{}
 	builder.WriteString("CREATE TABLE ")
-	builder.WriteString(tableNamePart(schema.TableName(sc), cluster))
+	builder.WriteString(tableNamePart(table.Name, cluster))
 	builder.WriteString(" (\n")
 	builder.WriteString("  ")
-	fields := sc.Fields()
-	for i, field := range fields {
-		definition, err := types.FieldDefinition(field)
+	for i, col := range table.Columns {
+		definition, err := types.FieldDefinition(col.ToArrowField())
 		if err != nil {
 			return "", err
 		}
 		builder.WriteString(definition)
-		if i < len(fields)-1 {
+		if i < len(table.Columns)-1 {
 			builder.WriteString(",\n  ")
 		}
 	}
 	builder.WriteString("\n) ENGINE = ")
 	builder.WriteString(engine.String())
-	if orderBy := sortKeys(sc); len(orderBy) > 0 {
+	if orderBy := sortKeys(table); len(orderBy) > 0 {
 		builder.WriteString(" ORDER BY (")
 		builder.WriteString(strings.Join(util.Sanitized(orderBy...), ", "))
 		builder.WriteString(")")
@@ -56,6 +54,6 @@ func CreateTable(sc *arrow.Schema, cluster string, engine *Engine) (string, erro
 	return builder.String(), nil
 }
 
-func DropTable(sc *arrow.Schema, cluster string) string {
-	return "DROP TABLE IF EXISTS " + tableNamePart(schema.TableName(sc), cluster)
+func DropTable(table *schema.Table, cluster string) string {
+	return "DROP TABLE IF EXISTS " + tableNamePart(table.Name, cluster)
 }