+++ plugins/destination/gremlin/client/migrate.go
@@ -3,10 +3,10 @@ package client
 import (
 	"context"
 
-	"github.com/cloudquery/plugin-sdk/v2/schema"
+	"github.com/cloudquery/plugin-sdk/v3/schema"
 )
 
 // Migrate tables. Like neo4j, gremlin does not have a schema, so this is a no-op.
-func (*Client) Migrate(ctx context.Context, tables schema.Schemas) error {
+func (*Client) Migrate(_ context.Context, _ schema.Tables) error {
 	return nil
 }