+++ plugins/destination/mongodb/client/deletestale.go
@@ -4,13 +4,13 @@ import (
 	"context"
 	"time"
 
-	"github.com/cloudquery/plugin-sdk/v2/schema"
+	"github.com/cloudquery/plugin-sdk/v3/schema"
 	"go.mongodb.org/mongo-driver/bson"
 )
 
-func (c *Client) DeleteStale(ctx context.Context, tables schema.Schemas, source string, syncTime time.Time) error {
+func (c *Client) DeleteStale(ctx context.Context, tables schema.Tables, source string, syncTime time.Time) error {
 	for _, table := range tables {
-		tableName := schema.TableName(table)
+		tableName := table.Name
 		// delete all records that are not in the source and are older than syncTime
 		if _, err := c.client.Database(c.pluginSpec.Database).Collection(tableName).DeleteMany(ctx, bson.M{
 			schema.CqSourceNameColumn.Name: source,