+++ plugins/destination/snowflake/client/deletestale.go
@@ -5,12 +5,12 @@ import (
 	"strings"
 	"time"
 
-	"github.com/cloudquery/plugin-sdk/v2/schema"
+	"github.com/cloudquery/plugin-sdk/v3/schema"
 )
 
-func (c *Client) DeleteStale(ctx context.Context, tables schema.Schemas, source string, syncTime time.Time) error {
+func (c *Client) DeleteStale(ctx context.Context, tables schema.Tables, source string, syncTime time.Time) error {
 	for _, table := range tables {
-		tableName := schema.TableName(table)
+		tableName := table.Name
 		var sb strings.Builder
 		sb.WriteString("delete from ")
 		sb.WriteString(tableName)